requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==10.2.0
PyJWT==2.6.0
pip==22.3.1
pytz==2024.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.4

